+++ examples/_archive/example_version_of_this_package.py
@@ -3,63 +3,31 @@
 #
 # File: example_version_of_this_package.py
 #
-# Part of ‘UNICORN Binance Rest API’
-# Project website: https://github.com/LUCIT-Systems-and-Development/unicorn-binance-websocket-api
-# Documentation: https://lucit-systems-and-development.github.io/unicorn-binance-websocket-api
-# PyPI: https://pypi.org/project/unicorn-binance-websocket-api/
+# Part of ‘UNICORN Binance REST API’
+# Project website: https://www.lucit.tech/unicorn-binance-rest-api.html
+# Github: https://github.com/LUCIT-Systems-and-Development/unicorn-binance-rest-api
+# Documentation: https://unicorn-binance-rest-api.docs.lucit.tech
+# PyPI: https://pypi.org/project/unicorn-binance-rest-api
+# LUCIT Online Shop: https://shop.lucit.services/software
+#
+# License: LSOSL - LUCIT Synergetic Open Source License
+# https://github.com/LUCIT-Systems-and-Development/unicorn-binance-rest-api/blob/master/LICENSE
 #
 # Author: LUCIT Systems and Development
 #
-# Copyright (c) 2019-2021, Oliver Zehentleitner
+# Copyright (c) 2022-2024, LUCIT Systems and Development (https://www.lucit.tech)
 # All rights reserved.
-#
-# Permission is hereby granted, free of charge, to any person obtaining a
-# copy of this software and associated documentation files (the
-# "Software"), to deal in the Software without restriction, including
-# without limitation the rights to use, copy, modify, merge, publish, dis-
-# tribute, sublicense, and/or sell copies of the Software, and to permit
-# persons to whom the Software is furnished to do so, subject to the fol-
-# lowing conditions:
-#
-# The above copyright notice and this permission notice shall be included
-# in all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-# OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABIL-
-# ITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT
-# SHALL THE AUTHOR BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-# WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
-# IN THE SOFTWARE.
 
-import unicorn_binance_rest_api
-import logging
-import os
+from unicorn_binance_rest_api import BinanceRestApiManager
 
-# https://docs.python.org/3/library/logging.html#logging-levels
-logging.getLogger("unicorn_binance_rest_api")
-logging.basicConfig(level=logging.DEBUG,
-                    filename=os.path.basename(__file__) + '.log',
-                    format="{asctime} [{levelname:8}] {process} {thread} {module}: {message}",
-                    style="{")
+with BinanceRestApiManager(warn_on_update=False) as ubra:
+    if ubra.is_update_available():
+        print(f"Please upgrade to {ubra.get_latest_version()} you are on {ubra.get_version()}")
+        latest_release_info = ubra.get_latest_release_info()
+        if latest_release_info:
+            print(f"Please download the latest release or run `pip install unicorn-binance-rest-api --upgrade`"
+                  f":\r\n\ttar: {latest_release_info['tarball_url']}\r\n\tzip: {latest_release_info['zipball_url']}\r\n"
+                  f"release info:\r\n{latest_release_info['body']}")
+    else:
+        print(ubra.get_version(), "is the latest version!")
 
-# create instance of BinanceRestApiManager
-# To use this library you need a valid UNICORN Binance Suite License: https://medium.lucit.tech/87b0088124a8
-ubra = unicorn_binance_rest_api.BinanceRestApiManager(exchange="binance.com")
-
-# get version of the used UNICORN Binance REST API package
-if ubra.is_update_availabe():
-    print("Please upgrade to " + ubra.get_latest_version() + ", you are on",
-          ubra.get_version())
-
-    latest_release_info = ubra.get_latest_release_info()
-    if latest_release_info:
-        print("Please download the latest release or run `pip install unicorn-binance-rest-api --upgrade`: ")
-        print("\ttar: " + latest_release_info["tarball_url"])
-        print("\tzip: " + latest_release_info["zipball_url"])
-        print("release info:")
-        print(latest_release_info["body"])
-else:
-    print(ubra.get_version(), "is the latest version!")
-
-ubra.stop_manager()