+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [How to upgrade to the latest version!](https://unicorn-binance-rest-api.docs.lucit.tech/readme.html#installation-and-upgrade)
 
 ## 2.5.1.dev (development stage/unreleased/unstable)
+### Fixed
+- `_save_used_weight()` - Exception: int() argument must be a string, a bytes-like object or a real number, not 'NoneType'
 
 ## 2.5.1
 ### Changed